Replace HTable with CuckooMap
- Simplify hash value manipulation

We do not use _mm_*_ps SIMD instructions any longer, so hash values do
not need to be non-NaN values when reinterpreted as a float.
Hash values should still be non-zero to detect empty slots in a bucket.

- Add CuckooMapWithVariableKeySize

CuckooMapWithVariableKeySize is a derived class of CuckooMap. It supports
custom hash/cmp function that takes an extra key size argument, which is
useful in ExactMatch and WildcardMatch